Refactor DelegateCollection & WeakDelegate
to share code between sync and async event source
